validasi_service/requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
redis
rq
flask_bcrypt
werkzeug

pyjwt

flask_mail
reportlab